pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "core"
version = "0.1.0"
description = "Prototyp ekstrakcji i transformacji"
authors = [
    {name = "Artur Podsiadły", email = "artpods56@gmail.com"}
]
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "numpy",
    "pandas",
    "torch",
    "transformers",
    "accelerate",
    "hydra-core",
    "wandb",
    "structlog",
    "rich",
    "diskcache==5.6.3",
    "geopandas",
    "jinja2",
    "mistralai",
    "omegaconf",
    "openai",
    "pillow",
    "pydantic",
    "pytesseract",
    "python-dotenv",
    "pyyaml",
    "rapidfuzz",
    "shapely",
    "streamlit",
    "openpyxl",
    "thefuzz",
    "tqdm",
    "typing_extensions",
    "lingua-language-detector>=2.0.0",
    "pymupdf",
    "psycopg2-binary>=2.9.10",
    "beartype>=0.21.0",
    "sqlalchemy>=2.0.43",
    "pytest>=8.4.2",
    "respx>=0.22.0",
    "llama-cpp-python>=0.3.16",
    "black>=25.9.0",
    "dagster>=1.11.15",
    "dagster-webserver>=1.11.15",
    "datasets>=4.2.0",
    "pyodbc>=5.3.0",
    "reportlab>=4.4.4",
]

[project.optional-dependencies]
# Used for language detection in the pipeline
lingua = [
    "lingua-language-detector>=2.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.dagster]
module_name = "orchestration.definitions"
code_location_name = "ai_osrodek"


[dependency-groups]
dev = [
    "basedpyright>=1.31.4",
]
